[proliferate] fix frontend release checks
